Jpa hibernate

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -6,6 +6,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,7 +15,7 @@ import com.algaworks.algafood.domain.repository.CozinhaRepository;
 
 @Component
 public class CozinhaRepositoryImpl implements CozinhaRepository {
-	
+
 	@PersistenceContext
 	private EntityManager entityManager;
 
@@ -29,13 +30,19 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
 	@Override
 	public Cozinha buscar(Long id) {
-		return entityManager.find(Cozinha.class, id);
+
+		Cozinha cozinha = entityManager.find(Cozinha.class, id);
+
+		if (cozinha == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		return cozinha;
 	}
 
 	@Transactional
 	@Override
 	public Cozinha salvar(Cozinha cozinha) {
-		
+
 		return entityManager.merge(cozinha);
 	}
 
@@ -45,8 +52,12 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
 		Cozinha cozinha = buscar(id);
 
+		if (cozinha == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+
 		entityManager.remove(cozinha);
-		
+
 	}
 
 }
